Commit 19
Created separate files for player and pokemon and  cpp/hpp
All classes except Professor Oak has its own cpp and hpp files
