+++ Pokemon/Pokemon/x64/Debug/Pokemon.tlog/link.secondary.1.tlog
@@ -1,2 +1,2 @@
-^C:\USERS\CASEY\DOCUMENTS\GITHUB\POKEMON\POKEMON\POKEMON\X64\DEBUG\MAIN.OBJ
+^C:\USERS\CASEY\DOCUMENTS\GITHUB\POKEMON\POKEMON\POKEMON\X64\DEBUG\MAIN.OBJ|C:\USERS\CASEY\DOCUMENTS\GITHUB\POKEMON\POKEMON\POKEMON\X64\DEBUG\PLAYER.OBJ|C:\USERS\CASEY\DOCUMENTS\GITHUB\POKEMON\POKEMON\POKEMON\X64\DEBUG\POKEMON.OBJ|C:\USERS\CASEY\DOCUMENTS\GITHUB\POKEMON\POKEMON\POKEMON\X64\DEBUG\UTILITY.OBJ
 C:\Users\Casey\Documents\GitHub\Pokemon\Pokemon\Pokemon\x64\Debug\Pokemon.ilk
