+++ Pokemon/main.cpp
@@ -1,113 +1,10 @@
 #include <string>
 #include <valarray>
-#include "PokemonType.hpp"
 #include "PokemonChoice.hpp"
+#include "PokemonType.hpp"
+#include "Pokemon.hpp"
 #include "Utility.hpp"
-
-
-class Pokemon {
-   public:
-    std::string Name;
-    PokemonType Type;
-    int Health;
-
-    Pokemon()
-    {
-
-        Name = "Pikachu";
-        Type = PokemonType::Electric;
-        Health = 10;
-        
-    }
-
-    Pokemon(std::string p_name, PokemonType p_type, int p_health)
-    {
-        
-        Name = p_name;
-        Type = p_type;
-        Health = p_health;
-        
-    }
-
-    
-    Pokemon(const Pokemon &other)
-    {
-
-        Name = other.Name;
-        Type = other.Type;
-        Health = other.Health;
-
-    }
-
-    ~Pokemon()
-    {
-     
-    }
-
-    void attack() { std::cout << Name << " Attacks with a powerful move" << "\n"; }
-};
-
-class Player
-{
-public:
-    std::string Name;
-    Pokemon chosenPokemon;
-    
-
-    
-    Player()
-    {
-        
-        Name = "Trainer";
-        chosenPokemon = Pokemon("Pikachu", PokemonType::Electric, 100);
-        
-    }
-
-    Player(std::string p_name, Pokemon p_Chosen)
-    {
-
-        Name = p_name;
-        chosenPokemon = p_Chosen;
-        
-        
-    }
-
-    void choosePokemon(int choice) {
-            switch ((PokemonChoice)choice)
-            {   
-            case PokemonChoice::Charmander:
-                MSG << "You have chosen Charmander!" << END;
-                chosenPokemon = Pokemon("Charmander", PokemonType::Fire, 100);
-                MSG << "Congratulations on picking your first Pokemon, Charmander! I'm sure you will become the best "
-                       "of friends! "<<std::cin.get() <<
-                       "It's time for you to step into the exciting world of Pokemon! "
-                       "Enjoy!" <<END;
-                break;
-            case PokemonChoice::Squirtle:
-                MSG << "You have chosen Squirtle!" << END;
-                chosenPokemon = Pokemon("Squirtle", PokemonType::Water, 100);
-                MSG << "Congratulations on picking your first Pokemon, Squirtle! I'm sure you will become the best of "
-                       "friends! " << std::cin.get() <<
-                       "It's time for you to step into the exciting world of Pokemon! "
-                       "Enjoy!"<<END;
-                break;
-            case PokemonChoice::Bulbasaur:
-                MSG << "You have chosen Bulbasaur!" << END;
-                chosenPokemon = Pokemon("Bulbasaur", PokemonType::Grass, 100);
-                MSG << "Congratulations on picking your first Pokemon, Bulbasaur! I'm sure you will become the best of "
-                       "friends! "
-               << std::cin.get() <<
-                       "It's time for you to step into the exciting world of Pokemon! "
-                       "Enjoy!" <<END;
-                break;
-            default:
-                chosenPokemon = Pokemon("Charmander", PokemonType::Fire, 100);
-                MSG << "You have not chosen a Pokemon on the list, let me choose for you! You have recieved Charmander "
-                       "since no one likes Bulbasaur!"<< END;
-            }
-      
-    }
-};
+#include "Player.hpp"
 
     class Professor
     {
